requirements.txt
alembic==1.16.5
BareNecessities==0.2.8
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mail==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
ngrok==1.5.1
packaging==25.0
pyngrok==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.43
text-unidecode==1.3
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
